M02A02 – Coerção de Tipo

+++ Modulo_02_Operadores/M02A02–Coercao_de_Tipo.c
@@ -6,7 +6,7 @@ setlocale(LC_ALL,"Portuguese");
 
     int x = 5;
     int y = 2;
-    int z = x / y;
+    float z = x / y;
 
-    printf("O resultado e %d",z);
+    printf("O resultado e %f",z);
 }
